---changes--- -fixed ratio problems with reading and writing of file
+++additions+++
+added export image method to project
+added get original pixelbuffer method to EditorController
+added get pixel array from drawtool

+++ src/main/java/org/editor/project/SettingsController.java
@@ -69,8 +69,8 @@ public class SettingsController implements Initializable, ControlScreen {
 
     public void handleCreateProject(ActionEvent event)  {
         try {
-            double projectWidth = Double.parseDouble(widthInput.getText());
-            double projectHeight = Double.parseDouble(heightInput.getText());
+            int projectWidth = Integer.parseInt(widthInput.getText());
+            int projectHeight = Integer.parseInt(heightInput.getText());
             if(!radioBackground.isSelected()){
                 if(cpBackgroundColor.getValue() != null){
                     project = new Project(projectWidth, projectHeight, cpBackgroundColor.getValue());
